Move await related classes to the old package
COSDK-480
