+++ action-core/api/action-core.api
@@ -39,8 +39,8 @@ public final class com/adyen/checkout/action/core/GenericActionConfiguration$Bui
 	public fun <init> (Ljava/util/Locale;Lcom/adyen/checkout/core/old/Environment;Ljava/lang/String;)V
 	public fun add3ds2ActionConfiguration (Lcom/adyen/checkout/adyen3ds2/Adyen3DS2Configuration;)Lcom/adyen/checkout/action/core/GenericActionConfiguration$Builder;
 	public synthetic fun add3ds2ActionConfiguration (Lcom/adyen/checkout/adyen3ds2/Adyen3DS2Configuration;)Ljava/lang/Object;
-	public fun addAwaitActionConfiguration (Lcom/adyen/checkout/await/AwaitConfiguration;)Lcom/adyen/checkout/action/core/GenericActionConfiguration$Builder;
-	public synthetic fun addAwaitActionConfiguration (Lcom/adyen/checkout/await/AwaitConfiguration;)Ljava/lang/Object;
+	public fun addAwaitActionConfiguration (Lcom/adyen/checkout/await/old/AwaitConfiguration;)Lcom/adyen/checkout/action/core/GenericActionConfiguration$Builder;
+	public synthetic fun addAwaitActionConfiguration (Lcom/adyen/checkout/await/old/AwaitConfiguration;)Ljava/lang/Object;
 	public fun addQRCodeActionConfiguration (Lcom/adyen/checkout/qrcode/QRCodeConfiguration;)Lcom/adyen/checkout/action/core/GenericActionConfiguration$Builder;
 	public synthetic fun addQRCodeActionConfiguration (Lcom/adyen/checkout/qrcode/QRCodeConfiguration;)Ljava/lang/Object;
 	public fun addRedirectActionConfiguration (Lcom/adyen/checkout/redirect/RedirectConfiguration;)Lcom/adyen/checkout/action/core/GenericActionConfiguration$Builder;
@@ -75,8 +75,8 @@ public abstract class com/adyen/checkout/action/core/internal/ActionHandlingPaym
 	public fun <init> (Ljava/util/Locale;Lcom/adyen/checkout/core/old/Environment;Ljava/lang/String;)V
 	public final fun add3ds2ActionConfiguration (Lcom/adyen/checkout/adyen3ds2/Adyen3DS2Configuration;)Lcom/adyen/checkout/components/core/internal/BaseConfigurationBuilder;
 	public synthetic fun add3ds2ActionConfiguration (Lcom/adyen/checkout/adyen3ds2/Adyen3DS2Configuration;)Ljava/lang/Object;
-	public final fun addAwaitActionConfiguration (Lcom/adyen/checkout/await/AwaitConfiguration;)Lcom/adyen/checkout/components/core/internal/BaseConfigurationBuilder;
-	public synthetic fun addAwaitActionConfiguration (Lcom/adyen/checkout/await/AwaitConfiguration;)Ljava/lang/Object;
+	public final fun addAwaitActionConfiguration (Lcom/adyen/checkout/await/old/AwaitConfiguration;)Lcom/adyen/checkout/components/core/internal/BaseConfigurationBuilder;
+	public synthetic fun addAwaitActionConfiguration (Lcom/adyen/checkout/await/old/AwaitConfiguration;)Ljava/lang/Object;
 	public final fun addQRCodeActionConfiguration (Lcom/adyen/checkout/qrcode/QRCodeConfiguration;)Lcom/adyen/checkout/components/core/internal/BaseConfigurationBuilder;
 	public synthetic fun addQRCodeActionConfiguration (Lcom/adyen/checkout/qrcode/QRCodeConfiguration;)Ljava/lang/Object;
 	public final fun addRedirectActionConfiguration (Lcom/adyen/checkout/redirect/RedirectConfiguration;)Lcom/adyen/checkout/components/core/internal/BaseConfigurationBuilder;
