+++ action-core/src/main/java/com/adyen/checkout/action/core/internal/ui/ActionDelegateProvider.kt
@@ -11,7 +11,7 @@ package com.adyen.checkout.action.core.internal.ui
 import android.app.Application
 import androidx.lifecycle.SavedStateHandle
 import com.adyen.checkout.adyen3ds2.internal.provider.Adyen3DS2ComponentProvider
-import com.adyen.checkout.await.internal.provider.AwaitComponentProvider
+import com.adyen.checkout.await.old.internal.provider.AwaitComponentProvider
 import com.adyen.checkout.components.core.CheckoutConfiguration
 import com.adyen.checkout.components.core.PaymentMethodTypes
 import com.adyen.checkout.components.core.action.Action
