+++ await/src/main/java/com/adyen/checkout/await/old/internal/ui/AwaitDelegate.kt
@@ -6,10 +6,10 @@
  * Created by oscars on 16/8/2022.
  */
 
-package com.adyen.checkout.await.internal.ui
+package com.adyen.checkout.await.old.internal.ui
 
 import androidx.annotation.RestrictTo
-import com.adyen.checkout.await.internal.ui.model.AwaitOutputData
+import com.adyen.checkout.await.old.internal.ui.model.AwaitOutputData
 import com.adyen.checkout.components.core.internal.ui.ActionDelegate
 import com.adyen.checkout.components.core.internal.ui.DetailsEmittingDelegate
 import com.adyen.checkout.components.core.internal.ui.RedirectableDelegate
