+++ await/src/main/java/com/adyen/checkout/await/old/internal/ui/AwaitViewProvider.kt
@@ -6,10 +6,10 @@
  * Created by josephj on 5/9/2022.
  */
 
-package com.adyen.checkout.await.internal.ui
+package com.adyen.checkout.await.old.internal.ui
 
 import android.content.Context
-import com.adyen.checkout.await.internal.ui.view.AwaitView
+import com.adyen.checkout.await.old.internal.ui.view.AwaitView
 import com.adyen.checkout.ui.core.internal.ui.ComponentView
 import com.adyen.checkout.ui.core.internal.ui.ComponentViewType
 import com.adyen.checkout.ui.core.internal.ui.ViewProvider
