+++ await/src/main/java/com/adyen/checkout/await/old/internal/ui/DefaultAwaitDelegate.kt
@@ -6,13 +6,13 @@
  * Created by oscars on 17/8/2022.
  */
 
-package com.adyen.checkout.await.internal.ui
+package com.adyen.checkout.await.old.internal.ui
 
 import android.app.Activity
 import androidx.annotation.VisibleForTesting
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.SavedStateHandle
-import com.adyen.checkout.await.internal.ui.model.AwaitOutputData
+import com.adyen.checkout.await.old.internal.ui.model.AwaitOutputData
 import com.adyen.checkout.components.core.ActionComponentData
 import com.adyen.checkout.components.core.action.Action
 import com.adyen.checkout.components.core.action.AwaitAction
