+++ await/src/main/java/com/adyen/checkout/await/old/internal/ui/view/AwaitView.kt
@@ -5,7 +5,7 @@
  *
  * Created by josephj on 31/8/2022.
  */
-package com.adyen.checkout.await.internal.ui.view
+package com.adyen.checkout.await.old.internal.ui.view
 
 import android.content.Context
 import android.util.AttributeSet
@@ -15,8 +15,8 @@ import android.widget.LinearLayout
 import androidx.annotation.StringRes
 import com.adyen.checkout.await.R
 import com.adyen.checkout.await.databinding.AwaitViewBinding
-import com.adyen.checkout.await.internal.ui.AwaitDelegate
-import com.adyen.checkout.await.internal.ui.model.AwaitOutputData
+import com.adyen.checkout.await.old.internal.ui.AwaitDelegate
+import com.adyen.checkout.await.old.internal.ui.model.AwaitOutputData
 import com.adyen.checkout.components.core.PaymentMethodTypes
 import com.adyen.checkout.components.core.internal.ui.ComponentDelegate
 import com.adyen.checkout.core.old.AdyenLogLevel
