+++ await/src/test/java/com/adyen/checkout/await/old/AwaitConfigurationTest.kt
@@ -1,4 +1,12 @@
-package com.adyen.checkout.await
+/*
+ * Copyright (c) 2025 Adyen N.V.
+ *
+ * This file is open source and available under the MIT license. See the LICENSE file for more info.
+ *
+ * Created by ararat on 11/7/2025.
+ */
+
+package com.adyen.checkout.await.old
 
 import com.adyen.checkout.components.core.Amount
 import com.adyen.checkout.components.core.AnalyticsConfiguration
